refactor: refactor ContinueWatchingRepositoryImplTest
